+++ headers/Wizard.h
@@ -10,8 +10,8 @@ using namespace std;
 namespace HE_Arc::RPG {
 class Wizard : public Hero {
 public:
-  Wizard(int strength = 10, int agility = 5, int intelligence = 1,
-         double hp = 20, string name = "unknown wizard", int mana = 10,
+  Wizard(int strength = 10, int agility = 5, int intelligence = 1, int hp = 20,
+         string name = "unknown wizard", int mana = 10,
          IObject *pObject = new Potion())
       : Hero(strength, agility, intelligence, hp, name, pObject), mana(mana){};
